TYPING Annotate the engine, and fix what the annotations found
Type annotations across OpenGLContext, with the defects they surfaced fixed
where they were found and a test written for each. Thirty-one new suites under
tests/unit cover the modules that had none: the debug helpers, the event
capture and release paths, the font registries and providers, the interpolators,
the atlas strip packer, the vector utilities, the GLFW and pygame key tables,
and the demo entry points under OpenGLContext/bin.

Behaviour that changed, and why:

- ``Context.DoInit`` puts deferred redraw back as it found it rather than
  switching it off. A backend's main loop defers for the whole session and
  wxContext runs DoInit from inside that loop, so clearing the flag left the
  session undeferred from its first paint.
- ``_MeshGPU.vao`` holds 0 rather than None for "no VAO", which is what GL
  itself means by the name and what ``glGenVertexArrays`` never returns.
- ``Context.setVSync`` answers False where there is no definition to record
  the preference on yet, rather than raising on the attribute.
- ``threading.Event.isSet`` is spelled ``is_set``.

+++ OpenGLContext/edit/mapview.py
@@ -184,7 +184,7 @@ class MapViewPlatform(ViewPlatform):
         super(MapViewPlatform, self).__init__(
             position=(view.centre[0], view.ceiling + STAND_OFF, view.centre[1]))
 
-    def setViewport(self, x: int, y: int) -> None:
+    def setViewport(self, x: float, y: float) -> None:
         self.viewport = (int(x), int(y))
 
     @property
